requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
aiofiles==24.1.0
boto3==1.35.0
celery[redis]==5.4.0
redis==5.2.0
httpx==0.28.0
Pillow==11.0.0
pytesseract==0.3.13
rapidfuzz==3.10.0
python-dateutil==2.9.0
pytz==2024.2
